Split up games into separate files
