+++ Marlin/src/lcd/menu/menu_main.cpp
@@ -46,6 +46,10 @@
   #include "../../feature/host_actions.h"
 #endif
 
+#if HAS_GAMES
+  #include "game/game.h"
+#endif
+
 #define MACHINE_CAN_STOP (EITHER(SDSUPPORT, HOST_PROMPT_SUPPORT) || defined(ACTION_ON_CANCEL))
 #define MACHINE_CAN_PAUSE (ANY(SDSUPPORT, HOST_PROMPT_SUPPORT, PARK_HEAD_ON_PAUSE) || defined(ACTION_ON_PAUSE))
 
@@ -136,16 +140,6 @@ void menu_led();
   #if SERVICE_INTERVAL_3 > 0
     void menu_service3();
   #endif
-#endif
-
-#if HAS_GAME_MENU
-  void menu_game();
-#elif ENABLED(MARLIN_BRICKOUT)
-  void lcd_goto_brickout();
-#elif ENABLED(MARLIN_INVADERS)
-  void lcd_goto_invaders();
-#elif ENABLED(MARLIN_SNAKE)
-  void lcd_goto_snake();
 #endif
 
 void menu_main() {
@@ -286,16 +280,18 @@ void menu_main() {
     #endif
   #endif
 
-  #if ANY(MARLIN_BRICKOUT, MARLIN_INVADERS, MARLIN_SNAKE)
+  #if ANY(MARLIN_BRICKOUT, MARLIN_INVADERS, MARLIN_SNAKE, MARLIN_MAZE)
     MENU_ITEM(submenu, "Game", (
       #if HAS_GAME_MENU
         menu_game
       #elif ENABLED(MARLIN_BRICKOUT)
-        lcd_goto_brickout
+        brickout.enter_game
       #elif ENABLED(MARLIN_INVADERS)
-        lcd_goto_invaders
+        invaders.enter_game
       #elif ENABLED(MARLIN_SNAKE)
-        lcd_goto_snake
+        snake.enter_game
+      #elif ENABLED(MARLIN_MAZE)
+        maze.enter_game
       #endif
     ));
   #endif
